refactor: extract SigIntHandler to _signals.py
